feat: remove target alias in trino merge
