+++ sqlglot/dialects/trino.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from sqlglot import exp
+from sqlglot.dialects.dialect import remove_target_from_merge
 from sqlglot.dialects.presto import Presto
 
 
@@ -11,6 +12,7 @@ class Trino(Presto):
         TRANSFORMS = {
             **Presto.Generator.TRANSFORMS,
             exp.ArraySum: lambda self, e: f"REDUCE({self.sql(e, 'this')}, 0, (acc, x) -> acc + x, acc -> acc)",
+            exp.Merge: remove_target_from_merge,
         }
 
     class Tokenizer(Presto.Tokenizer):
