+++ tests/dialects/test_postgres.py
@@ -740,6 +740,7 @@ class TestPostgres(Validator):
             """merge into x as x using (select id) as y on a = b WHEN matched then update set X."A" = y.b""",
             write={
                 "postgres": """MERGE INTO x AS x USING (SELECT id) AS y ON a = b WHEN MATCHED THEN UPDATE SET "A" = y.b""",
+                "trino": """MERGE INTO x AS x USING (SELECT id) AS y ON a = b WHEN MATCHED THEN UPDATE SET "A" = y.b""",
                 "snowflake": """MERGE INTO x AS x USING (SELECT id) AS y ON a = b WHEN MATCHED THEN UPDATE SET X."A" = y.b""",
             },
         )
